requirements.txt
instagrapi.git==2.1.3
pillow==11.1.0
pyinstaller==6.12.0